add by hyb for tx fullhash
