+++ blockchain/chain_test.go
@@ -478,6 +478,14 @@ func testProcQueryTxMsg(cfg *types.Chain33Config, t *testing.T, blockchain *bloc
 		assert.Nil(t, txProof.GetProofs())
 		assert.Nil(t, txProof.GetTxProofs()[0].GetProofs())
 	}
+
+	blockheight := block.Block.GetHeight()
+	if cfg.IsPara() {
+		blockheight = block.Block.GetMainHeight()
+	}
+	if cfg.IsFork(blockheight, "ForkRootHash") {
+		txhash = block.Block.Txs[txindex].FullHash()
+	}
 	//证明txproof的正确性,
 	if txProof.GetProofs() != nil { //ForkRootHash 之前的proof证明
 		brroothash := merkle.GetMerkleRootFromBranch(txProof.GetProofs(), txhash, uint32(txindex))
