+++ blockchain/filter_paratx.go
@@ -8,7 +8,6 @@ import (
 	"bytes"
 	"strings"
 
-	"github.com/33cn/chain33/common"
 	"github.com/33cn/chain33/common/merkle"
 	"github.com/33cn/chain33/types"
 )
@@ -194,74 +193,6 @@ func (chain *BlockChain) getChildChainProofs(height int64, blockHash []byte, tit
 		return branch, childHash, index
 	}
 	return nil, nil, 0
-}
-
-//获取指定交易在子链中的路径证明
-func (chain *BlockChain) getMultiLayerProofs(height int64, blockHash []byte, Txs []*types.Transaction, index int32) []*types.TxProof {
-
-	title, haveParaTx := types.GetParaExecTitleName(string(Txs[index].Execer))
-	if !haveParaTx {
-		title = types.MainChainName
-	}
-
-	replyparaTxs, err := chain.LoadParaTxByHeight(height, "", 0, 1)
-	if err != nil || 0 == len(replyparaTxs.Items) {
-		filterlog.Error("getMultiLayerProofs", "height", height, "blockHash", common.ToHex(blockHash), "err", err)
-		return nil
-	}
-
-	//获取本子链的第一笔交易的index值，以及最后一笔交易的index值
-	var startIndex int32
-	var childIndex uint32
-	var hashes [][]byte
-	var exist bool
-	var childHash []byte
-	var txCount int32
-	var proofs []*types.TxProof
-
-	for _, paratx := range replyparaTxs.Items {
-		if title == paratx.Title && bytes.Equal(blockHash, paratx.GetHash()) {
-			startIndex = paratx.GetStartIndex()
-			childIndex = paratx.ChildHashIndex
-			childHash = paratx.ChildHash
-			txCount = paratx.GetTxCount()
-			exist = true
-		}
-		hashes = append(hashes, paratx.ChildHash)
-	}
-	//只有主链的交易,没有其他平行链的交易
-	if len(replyparaTxs.Items) == 1 && startIndex == 0 && childIndex == 0 {
-		txProofs, _ := getTransactionProofs(Txs, index)
-		proof := types.TxProof{Proofs: txProofs, Index: uint32(index)}
-		proofs = append(proofs, &proof)
-		return proofs
-	}
-
-	endindex := startIndex + txCount
-	//不存在或者获取到的索引错误直接返回
-	if !exist || startIndex > index || endindex > int32(len(Txs)) || childIndex >= uint32(len(replyparaTxs.Items)) {
-		filterlog.Error("getMultiLayerProofs", "height", height, "blockHash", common.ToHex(blockHash), "exist", exist, "replyparaTxs", replyparaTxs)
-		filterlog.Error("getMultiLayerProofs", "startIndex", startIndex, "index", index, "childIndex", childIndex)
-		return nil
-	}
-
-	//主链和平行链交易都存在,获取本子链所有交易的hash值
-	var childHashes [][]byte
-	for i := startIndex; i < endindex; i++ {
-		childHashes = append(childHashes, Txs[i].Hash())
-	}
-	txOnChildIndex := index - startIndex
-
-	//计算单笔交易在子链中的路径证明
-	txOnChildBranch := merkle.GetMerkleBranch(childHashes, uint32(txOnChildIndex))
-	txproof := types.TxProof{Proofs: txOnChildBranch, Index: uint32(txOnChildIndex), RootHash: childHash}
-	proofs = append(proofs, &txproof)
-
-	//子链在整个区块中的路径证明
-	childBranch := merkle.GetMerkleBranch(hashes, childIndex)
-	childproof := types.TxProof{Proofs: childBranch, Index: childIndex}
-	proofs = append(proofs, &childproof)
-	return proofs
 }
 
 //LoadParaTxByTitle 通过title获取本平行链交易所在的区块高度信息前后翻页
